Added our own copy of urllib2 and modified it so that we can change the user-agent string. Added a config option for the user-agent.
Changed HTTPMonster to use our urllib2
Added dodgy HTML checking to URLThreads, so that google's fucked up broken anchor tags won't break the bot

Added a space in the __str__() for pluginreply!

Added the Airline plugin, which lets you lookup air carrier codes/names, and also lets you look up flight information if you give the carrier code and flight number, and optional date. ripped straight out of the PinkyBot java, and it is nearly as ugly.

Fixed a typo in  a comment(!) in news, and added some more LOG_DEBUG so I could see what news was doing

added "tell nick about whatever" to the SmartyPants, and added some escaping on the text that gets sent to the db in a listkeys/listvalues search.


git-svn-id: svn://svn.madcowdisease.org/mcd/trunk/blamehangle@1616 35f662cf-58f2-0310-8820-d172d141b225