Added "global" userlist (really it is only for ChatterGizmo) that lets you make the bot totally ignore someone, or flag someone as an "admin"
Added the ability to rehash config files while the bot is running. A user with the "admin" flag sends the bot a CTCP REHASH, and all config files are re-loaded, and every plugin (every Child, really) is sent a REQ_REHASH message. All plugin events are cleared, and every plugin is sent a PLUGIN_REGISTER message.


git-svn-id: svn://svn.madcowdisease.org/mcd/trunk/blamehangle@1663 35f662cf-58f2-0310-8820-d172d141b225