Overhaul the way we do threads for url grabbing and db queries.
Instead of spawning a new thread for each url/query, we now spawn a number of worker threads that are around for the life of the program. whenever a new url/query arrives, a worker thread will grab it and do the appropriate work, then go back to idling. This -should- remove the synchronization event handle bloat we were seeing in win32.

Added shutdown and rehash methods for HTTPMonster and DataMonkey.

Added missing calls to .close() on the StringIO objects used by several plugins.


git-svn-id: svn://svn.madcowdisease.org/mcd/trunk/blamehangle@1700 35f662cf-58f2-0310-8820-d172d141b225